requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
aiofiles==23.2.1
python-dotenv==1.0.0
aiohttp==3.9.1
